chore(Architecture): Setup Dependencies Injection and Bloc architecture
