+++ lib/ui/common/base_classes/blocs/BaseBloc.dart
@@ -1,0 +1,5 @@
+abstract class BaseBloc {
+
+  void dispose();
+
+}
