+++ pubspec.yaml
@@ -33,6 +33,9 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
+  ## dependencies injection
+  get_it: ^3.0.1
+
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
